# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/interface/add_list": {
            "post": {
                "description": "Create interface information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interface information"
                ],
                "summary": "Create interface information",
                "parameters": [
                    {
                        "description": "create interface request message",
                        "name": "addRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/interface/delete/{id}": {
            "get": {
                "description": "Delete interface information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interface information"
                ],
                "summary": "Delete interface information by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/admin/interface/update": {
            "post": {
                "description": "Update interface information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interface information"
                ],
                "summary": "Update interface information",
                "parameters": [
                    {
                        "description": "update interface request message",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
        "/api/interface_info/get_info/{id}": {
            "get": {
                "description": "Get interface information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interface information"
                ],
                "summary": "Get interface information by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
        "/api/interface_info/get_list": {
            "post": {
                "description": "Get interface information list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interface information"
                ],
                "summary": "Get interface information list",
                "parameters": [
                    {
                        "description": "get interface request message",
                        "name": "queryRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QueryInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.InfoResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "middlewares.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.AddInfoRequest": {
            "type": "object",
            "required": [
                "method",
                "status",
                "url"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string",
                    "maxLength": 256
                },
                "method": {
                    "description": "请求类型",
                    "type": "string",
                    "maxLength": 256,
                    "enum": [
                        "GET",
                        "POST",
                        "PUT",
                        "DELETE"
                    ]
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 256
                },
                "request_header": {
                    "description": "请求头",
                    "type": "string",
                    "maxLength": 8192
                },
                "response_header": {
                    "description": "响应头",
                    "type": "string",
                    "maxLength": 8192
                },
                "status": {
                    "description": "接口状态(0-关闭， 1-开启)",
                    "type": "integer",
                    "maximum": 10,
                    "enum": [
                        0,
                        1
                    ]
                },
                "url": {
                    "description": "接口地址",
                    "type": "string",
                    "maxLength": 512
                },
                "user_id": {
                    "description": "创建人",
                    "type": "integer"
                }
            }
        },
...
        "models.InfoResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "method": {
                    "description": "请求类型",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "request_header": {
                    "description": "请求头",
                    "type": "string"
                },
                "response_header": {
                    "description": "响应头",
                    "type": "string"
                },
                "status": {
                    "description": "接口状态(0-关闭， 1-开启)",
                    "type": "integer"
                },
                "url": {
                    "description": "接口地址",
                    "type": "string"
                },
                "user_id": {
                    "description": "创建人",
                    "type": "integer"
                }
            }
        },
...
        "models.QueryInfoRequest": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "maximum": 256
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 64
                }
            }
        },
...
        "models.UpdateInfoRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string",
                    "maxLength": 256
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "description": "请求类型",
                    "type": "string",
                    "maxLength": 256
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 256
                },
                "request_header": {
                    "description": "请求头",
                    "type": "string",
                    "maxLength": 8192
                },
                "response_header": {
                    "description": "响应头",
                    "type": "string",
                    "maxLength": 8192
                },
                "status": {
                    "description": "接口状态(0-关闭， 1-开启)",
                    "type": "integer",
                    "maximum": 10
                },
                "url": {
                    "description": "接口地址",
                    "type": "string",
                    "maxLength": 512
                },
                "user_id": {
                    "description": "创建人",
                    "type": "integer"
                }
            }
        },
...
    }
}